migrations/requirements.txt
# Migration Script Dependencies
# Install with: pip install -r migrations/requirements.txt

psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
